pyproject.toml
[project]
name = "strands"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore==0.1.5",
    "bedrock-agentcore-starter-toolkit==0.1.14",
    "strands-agents==1.9.1",
    "strands-agents-tools>=0.2.8",
]
